chore: move main index.js into lib directory
* add files section to package.json
* lint fix tests and fixtures
* lint fix readme
